Introduce new architecture.

+++ src/Window.cpp
@@ -13,23 +13,18 @@ namespace vzt
 	{
 	}
 
-	SurfaceHandler::~SurfaceHandler()
-	{
-		vkDestroySurfaceKHR(m_instance->VkHandle(), m_surface, nullptr);
-	}
+	SurfaceHandler::~SurfaceHandler() { vkDestroySurfaceKHR(m_instance->VkHandle(), m_surface, nullptr); }
 
-	Window::Window(
-	    std::string_view name, const uint32_t width, const uint32_t height,
-	    FrameBufferResizedCallback fbResizedCallback)
+	Window::Window(std::string_view name, const uint32_t width, const uint32_t height,
+	               FrameBufferResizedCallback fbResizedCallback)
 	    : m_width(width), m_height(height), m_fbResizedCallback(std::move(fbResizedCallback))
 	{
 		if (!glfwInit())
 			throw std::runtime_error("Failed to load glfw.");
 
 		glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
-		m_window = GLFWwindowPtr(
-		    glfwCreateWindow(m_width, m_height, std::string(name).c_str(), nullptr, nullptr),
-		    [](GLFWwindow *window) { glfwDestroyWindow(window); });
+		m_window = GLFWwindowPtr(glfwCreateWindow(m_width, m_height, std::string(name).c_str(), nullptr, nullptr),
+		                         [](GLFWwindow *window) { glfwDestroyWindow(window); });
 
 		glfwSetWindowUserPointer(m_window.get(), this);
 		glfwSetFramebufferSizeCallback(m_window.get(), [](GLFWwindow *window, int width, int height) {
@@ -38,12 +33,9 @@ namespace vzt
 		});
 	}
 
-	void Window::FrameBufferResized() const
-	{
-		m_fbResizedCallback();
-	}
+	void Window::FrameBufferResized() const { m_fbResizedCallback(); }
 
-	vzt::Size2D<int> Window::FrameBufferSize() const
+	vzt::Size2D<uint32_t> Window::FrameBufferSize() const
 	{
 		int frameBufferWidth = 0, frameBufferHeight = 0;
 		glfwGetFramebufferSize(m_window.get(), &frameBufferWidth, &frameBufferHeight);
@@ -52,7 +44,7 @@ namespace vzt
 			glfwGetFramebufferSize(m_window.get(), &frameBufferWidth, &frameBufferHeight);
 			glfwWaitEvents();
 		}
-		return {frameBufferWidth, frameBufferHeight};
+		return {static_cast<uint32_t>(frameBufferWidth), static_cast<uint32_t>(frameBufferHeight)};
 	}
 
 	VkSurfaceKHR Window::Surface(vzt::Instance *instance)
@@ -79,7 +71,7 @@ namespace vzt
 
 	std::vector<const char *> Window::VkExtensions() const
 	{
-		uint32_t glfwExtensionCount = 0;
+		uint32_t     glfwExtensionCount = 0;
 		const char **glfwExtensions;
 		glfwExtensions = glfwGetRequiredInstanceExtensions(&glfwExtensionCount);
